refactor: separate component input and label
