+++ tailwind.config.js
@@ -16,6 +16,7 @@ module.exports = {
         'c-white': '#f9f9f9',
         'c-white-broken': '#FAFDFB',
         'c-light-purple': '#95a0f4',
+        'c-light-purple-2': '#acb5f6',
         'c-purple': '#6777EF',
         'c-white-purple': '#fdfdff',
         'c-white-purple-2': '#e4e6fc',
